Bugs_but working kinda
