+++ src/components/Home.tsx
@@ -1,13 +1,15 @@
 import Classnames from 'classnames';
 import Home_module from './Home.module.scss';
+import { useNavigate } from 'react-router-dom';
 
 type Props = {
-    totalQuestions: number;
-    correctQuestions: number;    
+    //totalQuestions: number;
+    //correctQuestions: number;    
     onPress: (difficulty: string) => void;
 };
 
 function Home(props: Props) {
+    const navigate = useNavigate();
     return (
         <div className={Home_module.home}>
             <h1 className={Home_module.text}>
@@ -16,24 +18,24 @@ function Home(props: Props) {
 
             <div className={Home_module.inlineButtons}>
                 <button
-                    onClick={() =>{props.onPress("Daily")}} 
+                    onClick={() =>{props.onPress("Daily"), navigate('/daily')}} 
                     className={Home_module.optionButton}>
                     Daily
                 </button>
                 <button 
-                    onClick={() =>{props.onPress("Iconic")}}
+                    onClick={() =>{props.onPress("Iconic"), navigate('/results')}}
                     className={Home_module.optionButton}>
                     Iconic
                 </button>
                 <button 
-                    onClick={() =>{props.onPress("Hard")}}
+                    onClick={() =>{props.onPress("Hard"), navigate('/results')}}
                     className={Home_module.optionButton}>
                     Hard
                 </button>
             </div>
 
             <button
-                 onClick={() =>{props.onPress("Default")}}
+                 onClick={() =>{props.onPress("Default"), navigate('/results')}}
                 className={Home_module.startButton}
             >
                 Start Quiz
